+++ src/components/ResumeUpload/index.tsx
@@ -6,6 +6,7 @@ import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
 import { Label } from "@/components/ui/label"
 import { Card, CardContent } from "@/components/ui/card"
 import { Skeleton } from "@/components/ui/skeleton"
+import { formatFileSize } from "@/utils/formatFileSize"
 
 interface ResumeUploadProps {
   resumeType: 'PDF' | 'LINK';
@@ -23,14 +24,6 @@ interface ResumeUploadProps {
     size?: number;
   };
 }
-
-const formatFileSize = (bytes: number): string => {
-  if (bytes === 0) return '0 B';
-  const k = 1024;
-  const sizes = ['B', 'KB', 'MB', 'GB'];
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`;
-};
 
 export default function ResumeUpload({
   resumeType,
